sm_entry: simplify stack initialization out of atomic entry.
Since now we use a dedicated SSA region, instead of the stack, to store
interrupted execution context, we don't need to setup the stack within
the atomic region anymore.

Also note that the atomic entry section will already take care of
restoring the SSA (including sm_ssa_sp) if the enclave was interrupted.